Fixed issue where motors moved a little when controller plugged in

+++ Firmware/Motion_Engine/USBControllerUI.h
@@ -74,8 +74,8 @@ class USBControllerUI {
 
     uint8_t moveTimeMinutes;
     uint8_t moveTimeHours;
-    uint8_t accelTime;  // Percent of the time of the move that the dolly is accelerating (0-100)
-    uint8_t decelTime;  // Percent of the time of the move that the dolly is decelerating (0-100)
+    uint32_t accelTime;  // Percent of the time of the move that the dolly is accelerating (0-100)
+    uint32_t decelTime;  // Percent of the time of the move that the dolly is decelerating (0-100)
     uint32_t shotStartTime;
     uint32_t shotTimeMS;
     uint8_t isContinuous;
